steadfast-websocket/flattrade/requirements.txt
requests
websocket_client
./dist/NorenRestApiPy-0.0.20-py2.py3-none-any.whl
pandas
pyyaml
python-dotenv
flask
flask-cors
flask-socketio